Implement graph visualization and data set augmentation
Added methods to CGAN for graph visualization of a specific subnet.
Now the architecture of each subnet of CGAN can be printed to a pdf
for better understanding of how the pieces fit together. Added new
data set classes for augmenting the MNIST data set through traditional
augmentation, as well as CGAN augmentation. This will be used to compare
performance between various scenarios.