Finish cleaning up MNIST Notebook
Fixed a few last bugs, and while I didn't rerun the entire notebook yet,
it should be able to be re-run when I have more time in the near future.
The draw_cam function was half-baked from a commit this morning, and
I needed to finish writing out concluding remarks.