Implement dynamic convolutional architectures for netG, netD, and netE
In order to be able to handle images of unknown size from users,
it is important for the architectures to be initialized to the correct
settings. The strategy here took a few steps:
1. Assess input image size
2. Determine the factorization of the maximum cropped image size
such that it could be decomposed into a 3, 5, 7, 9, or 11 multiplied by
a factor of 2.
3. Crop image to this size.
4. Build an architecture of layers that perform the calculated up/down
sampling.
For example: If an image was size 30 x 30, this method would first
crop the image to 28 x 28, and then build layers such that the
upsampling looks like 7 --> 14 --> 28, etc.